Continue to capture SCP messages for previous ledger in database
Closes #3769

Upon closing a ledger, nodes record SCP messages received during the
ledger period in the `scphistory` database table. However, any messages
about that ledger heard *after* close were dropped. This change captures
more of those messages by modifying the `processExternalized` function
(which is responsible for recording SCP messages on ledger close) to
also record any new messages heard about the previous ledger.

+++ src/herder/HerderPersistenceImpl.cpp
@@ -37,7 +37,8 @@ HerderPersistenceImpl::~HerderPersistenceImpl()
 void
 HerderPersistenceImpl::saveSCPHistory(uint32_t seq,
                                       std::vector<SCPEnvelope> const& envs,
-                                      QuorumTracker::QuorumMap const& qmap)
+                                      QuorumTracker::QuorumMap const& qmap,
+                                      UnorderedSet<NodeID>& seenEnvs)
 {
     ZoneScoped;
     if (envs.empty())
@@ -50,7 +51,10 @@ HerderPersistenceImpl::saveSCPHistory(uint32_t seq,
 
     soci::transaction txscope(db.getSession());
 
+    if (seenEnvs.empty())
     {
+        // Clear information about ledger `seq` iff it's the first time we're
+        // recording info about this ledger
         auto prepClean = db.getPreparedStatement(
             "DELETE FROM scphistory WHERE ledgerseq =:l");
 
@@ -64,6 +68,12 @@ HerderPersistenceImpl::saveSCPHistory(uint32_t seq,
     }
     for (auto const& e : envs)
     {
+        if (seenEnvs.count(e.statement.nodeID))
+        {
+            continue;
+        }
+        seenEnvs.emplace(e.statement.nodeID);
+
         auto const& qHash =
             Slot::getCompanionQuorumSetHashFromStatement(e.statement);
         usedQSets.insert(
